+++ README.md
@@ -1,14 +1,14 @@
-# qiboicarusq
+# icarusq
 
-![Tests](https://github.com/qiboteam/qiboicarusq/workflows/Tests/badge.svg)
-[![codecov](https://codecov.io/gh/qiboteam/qiboicarusq/branch/main/graph/badge.svg?token=11UENAPBPH)](https://codecov.io/gh/qiboteam/qiboicarusq)
+![Tests](https://github.com/qiboteam/icarusq/workflows/Tests/badge.svg)
+[![codecov](https://codecov.io/gh/qiboteam/icarusq/branch/main/graph/badge.svg?token=11UENAPBPH)](https://codecov.io/gh/qiboteam/icarusq)
 ZENODO_BADGE
 
 This package provides the quantum hardware control implementation of the IcarusQ experiment at [CQT](https://www.quantumlah.org/) for [Qibo](https://github.com/qiboteam/qibo).
 
 ## Documentation
 
-The qiboicarusq backend documentation is available at [qiboicarusq.readthedocs.io](https://qiboicarusq.readthedocs.io/en/stable).
+The icarusq backend documentation is available at [icarusq.readthedocs.io](https://icarusq.readthedocs.io/en/stable).
 
 ## Citation policy
