+++ src/icarusq/calibration/procedure_batch.py
@@ -1,9 +1,6 @@
-import json
 import numpy as np
-from qiboicarusq import pulses, experiment, scheduler
-from qiboicarusq.calibration import fitting, tasks
-
-default_averaging = experiment.static.default_averaging
+from icarusq import pulses, experiment, scheduler
+from icarusq.calibration import fitting, tasks
 
 class Qubit:
     def __init__(self):
@@ -23,7 +20,7 @@ class Qubit:
         self.drive_channel = static_config["channel"][0]
 
 def _parse_result(raw_data, static_config):
-    final = experiment.static.sample_size / experiment.static.ADC_sampling_rate
+    final = experiment.static.ADC_length / experiment.static.ADC_sampling_rate
     step = 1 / experiment.static.ADC_sampling_rate
     ADC_time_array = np.arange(0, final, step)
     ADC_time_array = ADC_time_array[50:]
@@ -40,11 +37,23 @@ def _parse_result(raw_data, static_config):
 
     return [it, qt, ampl, phase]
 
-def _execute_pulse_sequences(pulse_sequences, static_config):
+def _parse_batch_result(raw_data, static_config):
+    steps = len(raw_data)
+    res = np.zeros((4, steps))
+    for i in range(steps):
+        it, qt, ampl, phase = _parse_result(raw_data[i], static_config)
+        res[0, i] = it
+        res[1, i] = qt
+        res[2, i] = ampl
+        res[3, i] = phase
+
+    return res
+
+def _execute_pulse_sequences(scheduler, pulse_sequences, static_config):
     steps = len(pulse_sequences)
     res = np.zeros((4, steps))
     for i in range(steps):
-        data = scheduler.execute_pulse_sequence(pulse_sequences[i], static_config).result()
+        data = scheduler.execute_pulse_sequence(pulse_sequences[i], experiment.static.default_averaging).result()
         it, qt, ampl, phase = _parse_result(data, static_config)
         res[0, i] = it
         res[1, i] = qt
@@ -64,18 +73,21 @@ def partial_qubit_calibration(static_config: dict, qubit: Qubit):
     channel = qubit.drive_channel
     amplitude = qubit.qubit_amplitude
     freq_sweep, seq = tasks.PulseSpectroscopy(freq_start, freq_end, amplitude, channel)
-    res = _execute_pulse_sequences(seq, static_config)
+    res = scheduler.execute_batch_sequence(seq, experiment.static.default_averaging).result()
+    res = _parse_batch_result(res, static_config)
     ampl_array = res[2]
     log["pulse"] = {
         "freq_sweep": freq_sweep.tolist(),
         "result": res.tolist()
     }
     freq = fitting.fit_pulse(ampl_array, freq_sweep)
+    log["freq"] = freq
     qubit.qubit_frequency = freq
 
     # Next, do Rabi oscillation to determine pi-pulse time
     time_sweep, seq = tasks.RabiTime(0, 600e-9, 3e-9, freq, amplitude, channel)
-    res = _execute_pulse_sequences(seq, static_config)
+    res = scheduler.execute_batch_sequence(seq, experiment.static.default_averaging).result()
+    res = _parse_batch_result(res, static_config)
     log["rabi"] = {
         "time_sweep": time_sweep.tolist(),
         "result": res.tolist()
@@ -88,12 +100,14 @@ def partial_qubit_calibration(static_config: dict, qubit: Qubit):
         "1": [res[0, idx_one], res[1, idx_one]]
     }
     freq_nyquist = freq - experiment.static.sampling_rate
-    qubit["rx"] = [pulses.BasicPulse(channel, 0, pi_pulse, amplitude, freq_nyquist, 0, pulses.Rectangular())]
-    qubit["ry"] = [pulses.BasicPulse(channel, 0, pi_pulse, amplitude, freq_nyquist, 90, pulses.Rectangular())]
+    qubit.pulses["rx"] = [pulses.BasicPulse(channel, 0, pi_pulse, amplitude, freq_nyquist, 0, pulses.Rectangular())]
+    qubit.pulses["ry"] = [pulses.BasicPulse(channel, 0, pi_pulse, amplitude, freq_nyquist, 90, pulses.Rectangular())]
+    log["pi-pulse"] = pi_pulse
 
     return qubit, log
 
 if __name__ == "__main__":
+    import json
     static_config = experiment.static.qubit_static_parameters[0]
     qb = Qubit()
     qb.load_from_staic_config(static_config)
