+++ src/icarusq/instruments/__init__.py
@@ -1,0 +1,6 @@
+from icarusq.instruments.awg import AWG
+from icarusq.instruments.attenuator import MCAttenuator
+from icarusq.instruments.quicsyn import QuicSyn
+from icarusq.instruments.rigol import RG
+from icarusq.instruments.instrument_controller import InstrumentController
+from icarusq.instruments.acquisition_controller import AcquisitionController
