+++ src/icarusq/instruments/attenuator.py
