+++ src/icarusq/instruments/quicsyn.py
