+++ src/icarusq/instruments/rigol.py
